refactor(Refactor of stop time normalization): Re-introduced batch processing
